+++ src/containers/home/stories/samples/modal-video/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
